requirements.txt
pyserial~=3.5
tweepy~=4.16.0
python-dotenv~=1.2.1
flask